Rename rewriter to pattern_rewriter
